new score file
