+++ rampwf/score_types/__init__.py
@@ -10,12 +10,13 @@ from .detection import (
     OSPA, SCP, DetectionPrecision, DetectionRecall, MADCenter, MADRadius,
     AverageDetectionPrecision, DetectionAveragePrecision)
 from .f1_above import F1Above
+from .generative_regression import (
+    NegativeLogLikelihoodReg, LikelihoodRatio,
+    NegativeLogLikelihoodRegDists, LikelihoodRatioDists)
 from .macro_averaged_recall import MacroAveragedRecall
 from .make_combined import MakeCombined
 from .mare import MARE
-from .negative_log_likelihood import NegativeLogLikelihood, \
-    NegativeLogLikelihoodReg, LikelihoodRatio, \
-    NegativeLogLikelihoodRegDists, LikelihoodRatioDists
+from .negative_log_likelihood import NegativeLogLikelihood
 from .normalized_gini import NormalizedGini
 from .normalized_rmse import NormalizedRMSE
 from .relative_rmse import RelativeRMSE
